# feat!: coverage-c8 to use V8 profiler directly instead of `NODE_V8_COVERAGE`