--- a/packages/coverage-c8/package.json
+++ b/packages/coverage-c8/package.json
@@ -42,7 +42,7 @@
     "prepublishOnly": "pnpm build"
   },
   "peerDependencies": {
-    "vitest": ">=0.28.0 <1"
+    "vitest": ">=0.29.0 <1"
   },
   "dependencies": {
     "c8": "^7.12.0",