--- a/packages/coverage-c8/src/provider.ts
+++ b/packages/coverage-c8/src/provider.ts
@@ -1,14 +1,13 @@
 import { existsSync, promises as fs } from 'fs'
 import _url from 'url'
 import type { Profiler } from 'inspector'
-import { takeCoverage } from 'v8'
 import { extname, resolve } from 'pathe'
 import c from 'picocolors'
 import { provider } from 'std-env'
 import type { RawSourceMap } from 'vite-node'
 import { coverageConfigDefaults } from 'vitest/config'
 // eslint-disable-next-line no-restricted-imports
-import type { CoverageC8Options, CoverageProvider, ReportContext, ResolvedCoverageOptions } from 'vitest'
+import type { AfterSuiteRunMeta, CoverageC8Options, CoverageProvider, ReportContext, ResolvedCoverageOptions } from 'vitest'
 import type { Vitest } from 'vitest/node'
 import type { Report } from 'c8'
 // @ts-expect-error missing types
@@ -16,15 +15,14 @@ import createReport from 'c8/lib/report.js'
 // @ts-expect-error missing types
 import { checkCoverages } from 'c8/lib/commands/check-coverage.js'
 
-type Options =
-  & ResolvedCoverageOptions<'c8'>
-  & { tempDirectory: string }
+type Options = ResolvedCoverageOptions<'c8'>
 
 export class C8CoverageProvider implements CoverageProvider {
   name = 'c8'
 
   ctx!: Vitest
   options!: Options
+  coverages: Profiler.TakePreciseCoverageReturnType[] = []
 
   initialize(ctx: Vitest) {
     this.ctx = ctx
@@ -35,25 +33,18 @@ export class C8CoverageProvider implements CoverageProvider {
     return this.options
   }
 
-  onBeforeFilesRun() {
-    process.env.NODE_V8_COVERAGE ||= this.options.tempDirectory
-  }
-
   async clean(clean = true) {
     if (clean && existsSync(this.options.reportsDirectory))
       await fs.rm(this.options.reportsDirectory, { recursive: true, force: true, maxRetries: 10 })
 
-    if (!existsSync(this.options.tempDirectory))
-      await fs.mkdir(this.options.tempDirectory, { recursive: true })
+    this.coverages = []
   }
 
-  onAfterSuiteRun() {
-    takeCoverage()
+  onAfterSuiteRun({ coverage }: AfterSuiteRunMeta) {
+    this.coverages.push(coverage as Profiler.TakePreciseCoverageReturnType)
   }
 
   async reportCoverage({ allTestsRun }: ReportContext = {}) {
-    takeCoverage()
-
     if (provider === 'stackblitz')
       this.ctx.logger.log(c.blue(' % ') + c.yellow('@vitest/coverage-c8 does not work on Stackblitz. Report will be empty.'))
 
@@ -64,6 +55,9 @@ export class C8CoverageProvider implements CoverageProvider {
 
     const report = createReport(options)
 
+    // Overwrite C8's loader as results are in memory instead of file system
+    report._loadReports = () => this.coverages
+
     interface MapAndSource { map: RawSourceMap; source: string | undefined }
     type SourceMapMeta = { url: string; filepath: string } & MapAndSource
 
@@ -73,7 +67,7 @@ export class C8CoverageProvider implements CoverageProvider {
 
     const entries = Array
       .from(this.ctx.vitenode.fetchCache.entries())
-      .filter(i => !i[0].includes('/node_modules/'))
+      .filter(entry => report._shouldInstrument(entry[0]))
       .map(([file, { result }]) => {
         if (!result.map)
           return null
@@ -153,12 +147,6 @@ export class C8CoverageProvider implements CoverageProvider {
 
     await report.run()
     await checkCoverages(options, report)
-
-    // Note that this will only clean up the V8 reports generated so far.
-    // There will still be a temp directory with some reports when vitest exists,
-    // but at least it will only contain reports of vitest's internal functions.
-    if (existsSync(this.options.tempDirectory))
-      await fs.rm(this.options.tempDirectory, { recursive: true, force: true, maxRetries: 10 })
   }
 }
 
@@ -178,7 +166,6 @@ function resolveC8Options(options: CoverageC8Options, root: string): Options {
 
     // Resolved fields
     provider: 'c8',
-    tempDirectory: process.env.NODE_V8_COVERAGE || resolve(reportsDirectory, 'tmp'),
     reporter: Array.isArray(reporter) ? reporter : [reporter],
     reportsDirectory,
   }