--- a/test/coverage-test/test/configuration-options.test-d.ts
+++ b/test/coverage-test/test/configuration-options.test-d.ts
@@ -116,6 +116,8 @@ test('provider module', () => {
       }
     },
     takeCoverage() {},
+    startCoverage() {},
+    stopCoverage() {},
   })
 })