… test/coverage-test/coverage-report-tests/__snapshots__/custom.report.test.ts.snap