Breaking changes:
- `vitest` and `@vitest/coverage-c8` versions must match as the `CoverageProviderModule` introduces new APIs and c8-provider starts to use these. 
- Removes `CoverageProvider.onBeforeFilesRun`. Custom coverage providers should use  `CoverageProviderModule.startCoverage` instead.

Currently Vitest is using [`NODE_V8_COVERAGE`](https://github.com/nodejs/node/blob/main/doc/api/cli.md#node_v8_coveragedir) environment variable to generate V8 coverage reports for `@vitest/coverage-c8`. We've used this environment variable to trigger Node to start writing JSON reports to `<reports-dir>/tmp`. 

These reports contain coverage results of all run code, including Node's internal `node:<module>` and unnecessary third party `**/node_modules/**` packages. In addition the results contain [`source-map-cache` information](https://github.com/nodejs/node/blob/main/doc/api/cli.md#source-map-cache) which is completely irrelevant for Vitest. There is so much data on these JSON reports. In #2108 user reported **435 MB**  `coverage/tmp` directory on CI. 

Also using `NODE_V8_COVERAGE` with `worker_threads` has some weird features: https://github.com/nodejs/node/issues/46378. I think it's meant more for running a single script instead of to be utilized by test runners. 

This PR replaces `process.env.NODE_V8_COVERAGE` usage by controlling V8 Profiler directly using `node:inspect`. We don't enable unncessary `source-map-cache` and filter the ScriptCoverage results early so the amount of stored data is reduced a lot. Results are stored in memory instead of file system - exactly as `@vitest/coverage-istanbul` does. This reduces amount of file system handles. Since we now know what files are in the report, we can reduce the amount of Vite source map handling in [`provider.ts`](https://github.com/vitest-dev/vitest/blob/4fc492c61d6e01bc34bc807a6b57197a1914f823/packages/coverage-c8/src/provider.ts#L74-L76). 

This should help with flaky coverage test on CI since workers no more write to `<reports-dir>/tmp` when they exit.