@AriPerkkio 2023-02-01
Minimal example for testing Profiler's coverage commands: https://gist.github.com/AriPerkkio/c2df034cd71b0dfcc934c477758c20ee