@AriPerkkio 2023-02-03
I did some manual regression testing with `vuejs/core` repository as these changes are quite critical. This seems to make report generation a lot faster now that amount of data is reduced, 3600ms -> 400ms:

```
Main branch:
provider.ts :: reportCoverage: 3.667s
provider.ts :: reportCoverage: 3.526s
provider.ts :: reportCoverage: 3.699s

PR branch:
provider.ts :: reportCoverage: 405.988ms
provider.ts :: reportCoverage: 399.793ms
provider.ts :: reportCoverage: 449.856ms
```


Coverage reports are almost identical. Branch count is off by one but after checking that one manually it doesn't make sense in either case.

```
Main:
=============================== Coverage summary ===============================
Statements   : 94.27% ( 39675/42086 )
Branches     : 92.07% ( 8987/9760 )
Functions    : 91.73% ( 2053/2238 )
Lines        : 94.27% ( 39675/42086 )
================================================================================

PR branch:
=============================== Coverage summary ===============================
Statements   : 94.27% ( 39675/42086 )
Branches     : 92.07% ( 8986/9759 )
Functions    : 91.73% ( 2053/2238 )
Lines        : 94.27% ( 39675/42086 )
================================================================================
```