@sheremet-va 2023-02-03
> This seems to make report generation a lot faster now that amount of data is reduced, 3600ms -> 400ms:

These are great improvements 😱 👍🏻