@AriPerkkio 2023-02-08
❗ Some manual resolving will be required if #2817 is merged first. Custom provider test should be updated to use the new `CoverageProviderModule`'s APIs as well.