@AriPerkkio 2023-02-13
Rebase done, most conflicts were in `integrations/coverage.ts`. Ubuntu+Node18 CI is failing on unrelated test. 
The `test/coverage-test/` tests pass and manual regression with `vuejs/core` gives good results:

<details>

```
# PR branch

 % Coverage report from c8

=============================== Coverage summary ===============================
Statements   : 94.27% ( 39675/42086 )
Branches     : 92.07% ( 8986/9759 )
Functions    : 91.73% ( 2053/2238 )
Lines        : 94.27% ( 39675/42086 )
================================================================================
reportCoverage: 418.102ms


# Main branch
 % Coverage report from c8

=============================== Coverage summary ===============================
Statements   : 94.27% ( 39675/42086 )
Branches     : 92.07% ( 8987/9760 )
Functions    : 91.73% ( 2053/2238 )
Lines        : 94.27% ( 39675/42086 )
================================================================================
reportCoverage: 3.828s
```

</details>